+++ daemon/include/MessageReceiver.hpp
@@ -28,7 +28,7 @@ public:
 	MessageReceiver();
 	~MessageReceiver();
 
-	virtual void onMessageReceived(quint8 id, const QByteArray &data) = 0;
+	virtual void onMessageReceived(quint16 id, const QByteArray &data) = 0;
 
 protected:
 	void handleIncomingData(const QByteArray &readData);
@@ -36,6 +36,6 @@ protected:
 	RxStatus m_rxStatus = MSG_RX_MAGIC;
 	quint16 m_rxByteCount = 0;
 	quint16 m_rxMsgSize = 0;
-	quint8 m_rxMsgID = 0;
+	quint16 m_rxMsgID = 0;
 	QByteArray m_rxBuffer;
 };
